% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBL.R
\docType{package}
\name{LBL-package}
\alias{LBL-package}
\title{LBL: Logistic Bayesian Lasso for Detecting Rare (or Common) Haplotype Association}
\description{
LBL uses the Bayesian LASSO framework to detect association between
a phenotype and haplotypes given the (unphased) genotypes of individuals.
}
\details{
LBL uses Bayesian Lasso to detect rare haplotypes that are associated with common diseases. 
The current implementation considers dichotomous traits. A future release will include quantitative and survival traits.
LBL is capable of handling different study designs: this version of the software is capable of handling independent cases and controls, case-parent trios and a mixture of both (provided
that the family data is independent of the case-control data).

A function from the hapassoc package is first implemented to acquire all compatible haplotypes. 
The posterior samples are then obtained via Markov Chain Monte Carlo (MCMC)
algorithm and inference on the parameters of interest can be carried out
(Bayes Factor, Credible Interval, etc.) based on these posterior samples.
}
\section{Functions}{

\code{\link{LBL}}:  MCMC algorithm to obtain posterior samples for independent case-control data.
\code{\link{famLBL}}:  MCMC algorithm to obtain posterior samples for case-parent trio data.
\code{\link{cLBL}}:  MCMC algorithm to obtain posterior samples for combined data.

\code{\link{LBL_summary}} provides model summary (in the form of list) based on posterior samples.
\code{\link{print_LBL_summary}} prints model summary in a user-friendly format from the list result of \code{\link{LBL_summary}}.
}

\references{
Biswas S. and Lin S. (2012). Logistic Bayesian LASSO for identifying
  association with rare haplotypes and application to age-related macular
  degeneration. \emph{Biometrics}, 68(2): 587-97.

Wang, M. and Lin, S. (2014). FamLBL: detecting rare haplotype disease association
  based on common SNPs using case-parent triads. \emph{Bioinformatics}, 30(18), 2611-2618.

Zhou, X., Wang, M., and Lin, S. (2019). cLBL: Combined logistic Bayesian LASSO for 
detecting rare associated haplotypes using independent case, control and family trio 
data. \emph{Manuscript}.
}
\author{
Swati Biswas, Meng Wang, Xiaofei Zhou, Han Zhang, Shuang Xia, Yuan Zhang, and Shili Lin <shili@stat.osu.edu>
}
