% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBL-package.R, R/LBLcac.R
\docType{package}
\name{LBL}
\alias{LBL}
\alias{LBL-package}
\alias{LBL}
\title{LBL: Logistic Bayesian Lasso for detecting Rare (or Common) Haplotype Association}
\usage{
LBL(data.cac, baseline = "missing", a = 15, b = 15, start.beta = 0.01,
  lambda = 1, D = 0, seed = NULL, burn.in = 10000, num.it = 40000,
  summary = T, e = 0.1, ci.level = 0.95)
}
\arguments{
\item{data.cac}{Input data. data.cac should be either a data frame or a matrix,
consisting of "n" rows and 6+2*p
 columns, where n is the number of cases and controls, and p
 is the number of SNPs.  The data should be in standard pedigree format, with
 the first 6 columns representing the family ID, individual ID, father ID,
 mother ID, sex, and affection status. The other 2*p columns are genotype
 data in allelic format, with each allele of a SNP taking up one column.
 An example can be found in this package under the name \code{"cac"}. For
 more information about the format, type \code{"?cac"} into R, or see "Linkage Format"
 section of \url{https://www.broadinstitute.org/haploview/input-file-formats}.
 Note that since these are independent case-control data, the father ID and mother ID
  are missing (coded as 0) and each individual has an unique family ID.}

\item{baseline}{Haplotype to be used for baseline coding; default is the most
frequent haplotype according to the initial haplotype frequency estimates.
This argument should be a character, starting with an h and followed by the
SNPs at each marker locus, for example, if the desired baseline haplotype is 0 1 1 0 0,
then baseline should be coded as "h01100".}

\item{a}{First hyperparameter of the prior for regression coefficients,
\strong{\eqn{\beta}}. The prior variance of \eqn{\beta} is 2/\eqn{\lambda^2} and \eqn{\lambda} has Gamma(a,b)
prior. The Gamma prior parameters a and b are such that the mean and variance of
the Gamma distribution are \eqn{a/b} and \eqn{a/b^2}. The default value of a is 15.}

\item{b}{Second hyperparameter of the Gamma(a,b) distribution described above; default
is 15.}

\item{start.beta}{Starting value of all regression coefficients, \strong{\eqn{\beta}};
default is 0.01.}

\item{lambda}{Starting value of the \eqn{\lambda} parameter described above;
default is 1.}

\item{D}{Starting value of the D parameter, which is the within-population
inbreeding coefficient; default is 0.}

\item{seed}{Seed to be used for the MCMC in Bayesian Lasso; default is a
random seed. If exact same results need to be reproduced, seed should be
fixed to the same number.}

\item{burn.in}{Burn-in period of the MCMC sampling scheme; default is 10000.}

\item{num.it}{Total number of MCMC iterations including burn-in; default is
40000.}

\item{summary}{Logical; if \code{TRUE}, LBL will return a summary of the analysis in the form of a list. If \code{FALSE}, LBL will return the posterior samples for all parameters.
Default is set to be TRUE.}

\item{e}{A (small) number \eqn{\epsilon} in the null hypothesis of no association,
\eqn{H_0: |\beta| \le \epsilon}. The default is 0.1. Changing e from the default of 0.1 may necessitate choosing a
different threshold for Bayes Factor (one of the outputs) to infer
association. Only used if \code{summary = TRUE}.}

\item{ci.level}{Credible probability. The probability that the true value of \eqn{beta} will
be within the credible interval. Default is 0.95, which corresponds to a 95\% posterior credible interval. Only used if \code{summary = TRUE}.}
}
\value{
If \code{summary = FALSE}, return a list with the following components:
\describe{
\item{haplotypes}{The list of haplotypes used in the analysis. The last column is the reference haplotype.}

\item{beta}{Posterior samples of betas stored in a matrix.}

\item{lambda}{a vector of (num.it-burn.in) posterior samples of lambda.}

\item{freq}{Posterior samples of the frequencies of haplotypes stored in a matrix format, in the same order as haplotypes.}

\item{init.freq}{The haplotype distribution used to initiate the MCMC.}

}

If \code{summary = TRUE}, return the result of LBL_summary.
 For details, see the description of the \code{?LBL_summary} function.
}
\description{
LBL uses the Bayesian LASSO framework to detect association between
a phenotype and haplotypes given the (unphased) genotypes of individuals.

\code{LBL} is a Bayesian LASSO method developed to detect association between
common/rare haplotypes and dichotomous disease phenotype, based on MCMC algorithm.
This function will handle independent cases and controls study design.
For other types of study designs, see \code{\link{famLBL}} and
\code{\link{cLBL}}. This function takes standard pedigree format as input with an individual's
genotypes, phenotype and familiar relationships, uses
an algorithm developed by Burkett \emph{et al}. to impute possible haplotype configurations
(for more detail, see \href{https://www.rdocumentation.org/packages/hapassoc/versions/0.5-1/topics/pre.hapassoc}{pre.hapassoc} from \pkg{hapassoc} package).
The input does not allow missing observations, and therefore subjects with
missing data are removed. This function returns an object containing summary statistics
for all haplotypes that are consistent with the input phenotype data.
}
\details{
LBL uses Bayesian Lasso to detect rare haplotypes that are associated with common diseases. 
The current implementation considers dichotomous traits. A future release will include quantitative and survival traits.
LBL is capable of handling different study designs: this version of the software is capable of handling independent cases and controls, case-parent trios and a mixture of both (provided
that the family data is independent of the case-control data). LBL uses algorithms from \href{https://www.rdocumentation.org/packages/hapassoc/versions/0.5-1/topics/pre.hapassoc}{pre.hapassoc}
function from \pkg{hapassoc} package to obtain all possible haplotype pairs compatible with an indivdiual's set of genotypes.

The posterior samples are obtained via Markov Chain Monte Carlo (MCMC)
algorithm and inference on the parameters of interest can be carried out
(Bayes Factor, Credible Interval, etc.) based on the posterior samples.
}
\section{Functions}{

\code{\link{LBL}}:  MCMC algorithm to obtain posterior samples for independent case-control data.
\code{\link{famLBL}}:  MCMC algorithm to obtain posterior samples for case-parent trio data.
\code{\link{cLBL}}:  MCMC algorithm to obtain posterior samples for combined data.

\code{\link{LBL_summary}} provides model summary (in the form of list) based on posterior samples.
\code{\link{print_LBL_summary}} prints model summary in a user-friendly format from the list result of \code{\link{LBL_summary}}.
}

\examples{
 data(cac)
 cac.obj<-LBL(cac)
 cac.obj
 print_LBL_summary(cac.obj)


}
\references{
Biswas S, Lin S (2012). Logistic Bayesian LASSO for Identifying
  Association with Rare Haplotypes and Application to Age-related Macular
  Degeneration. Biometrics, 68(2): 587-97.

Wang, M., & Lin, S. (2014). FamLBL: detecting rare haplotype disease association
based on common SNPs using case-parent triads. Bioinformatics, 30(18), 2611-2618.

Burkett, K., Graham, k. and McNeney, B. (2006). "hapassoc: Software for likelihood
inference of trait associations with SNP haplotypes and other attributes."
Journal of Statistical Software 16(2): 1-19.
}
\seealso{
\code{\link{famLBL}}, \code{\link{cLBL}}, \code{\link{LBL_summary}}.
}
\author{
Swati Biswas, Meng Wang, Xiaofei Zhou, Han Zhang, Shuang Xia, Yuan Zhang, and Shili Lin <shili@stat.osu.edu>
}
