% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBL_summary.R
\name{LBL_summary}
\alias{LBL_summary}
\title{Posterior Inference for LBL}
\usage{
LBL_summary(output, a = 15, b = 15, e = 0.1, ci.level = 0.95)
}
\arguments{
\item{output}{an obect returned by LBL with `summary=FALSE`.}

\item{a}{first hyperparameter of the prior for regression coefficients,
\eqn{\beta}. The prior variance of \eqn{\beta} is 2/\eqn{\lambda^2} and \eqn{\lambda} has Gamma(a,b)
prior. The Gamma prior parameters a and b are such that the mean and variance of
the Gamma distribution are \eqn{a/b} and \eqn{a/b^2}. The default value of a is 15.}

\item{b}{b parameter of the Gamma(a,b) distribution described above; default
is 15.}

\item{e}{a (small) number \eqn{\epsilon} in the null hypothesis of no association,
\eqn{H_0: |\beta| \le \epsilon}. The default is 0.1. Changing e from default of 0.1 may need choosing a
different threshold for Bayes Factor (one of the outputs) to infer
association.}

\item{ci.level}{Confidence level. The probability that the true value of \eqn{beta} will
be within the confidence interval. Default is 0.95 which corresponds to a 95\% posterior confidence interval. Only used if `summary = TRUE`.}
}
\value{
A list with the following components:
\describe{
\item{haplotypes}{The list of haplotypes used in the analysis.}

\item{OR}{Posterior mean of odds ratio.}

\item{OR.CI}{95\% posterior credible sets for the ORs.}

\item{BF}{Bayes Factor estimates based on posterior samples.
 if the posterior samples are all greater than e, then BF is set to be 999.}

\item{Significance}{Denotes the significance level of a haplotype. “*+” denotes a risk haplotype (\eqn{theta > 0},
 confidence interval of \eqn{theta} does not include 0, and BF >2). “*-” denotes a protective haplotype (\eqn{theta < 0},
 confidence interval of \eqn{theta} does not include 0, and BF >2). No labels indicates the haplotype is not sifnigicant.}

}
}
\description{
\code{LBL_summary} Provides inferences based on the posterior samples. Speficially, this function will provide posterior means,
confidence intervals and Bayes Factors (BF) estimates for the haplotypic effect coefficients.
}
\references{
Biswas S, Lin S (2012). Logistic Bayesian LASSO for Identifying
  Association with Rare Haplotypes and Application to Age-related Macular
  Degeneration. Biometrics, 68(2): 587-97.
}
