% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBLcac.R
\name{LBLcac}
\alias{LBLcac}
\title{Logistic Bayesian Lasso for detecting Rare (and Common) Haplotyptic Association with
Population Based Designs}
\usage{
LBLcac(data.cac, baseline = "missing", a = 15, b = 15,
  start.beta = 0.01, lambda = 1, D = 0, seed = NULL, burn.in = 10000,
  num.it = 40000, summary = T, e = 0.1, ci.level = 0.95)
}
\arguments{
\item{data.cac}{Input data. data.cac should be either a data frame or a matrix,
consist of "n" rows and 6+2*p
 columns, where n is the number of individuals of case-parent trios, and p
 is the number of SNPs.  The data should be standard pedigree format, with
 the first 6 columns representing the family ID, individual ID, father ID,
 mother ID, sex, and affection status. The other 2*p columns are genotype
 data in allelic format, with each allele of a SNP taking up one column.
 An example can be found in this package under the name \code{"cac"}. For
 more information about the format, type \code{"?cac"} into R, or see "Linkage Format"
 section of \url{https://www.broadinstitute.org/haploview/input-file-formats}.}

\item{baseline}{haplotype to be used for baseline coding; default is the most
frequent haplotype according to the initial haplotype frequency estimates.
This argument should be a character, starting with an h and followed by the
SNPs at each marker locus, for example, if the desired baseline haplotype is 0 1 1 0 0,
then baseline should be coded as "h01100".}

\item{a}{first hyperparameter of the prior for regression coefficients,
\eqn{\beta}. The prior variance of \eqn{\beta} is 2/\eqn{\lambda^2} and \eqn{\lambda} has Gamma(a,b)
prior. The Gamma prior parameters a and b are such that the mean and variance of
the Gamma distribution are \eqn{a/b} and \eqn{a/b^2}. The default value of a is 15.}

\item{b}{b parameter of the Gamma(a,b) distribution described above; default
is 15.}

\item{start.beta}{starting value of all regression coefficients, \eqn{\beta};
default is 0.01.}

\item{lambda}{starting value of the \eqn{\lambda} parameter described above;
default is 1.}

\item{D}{starting value of the D parameter, which is the within-population
inbreeding coefficient; default is 0.}

\item{seed}{the seed to be used for the MCMC in Bayesian Lasso; default is a
random seed. If exactly same results need to be reproduced, seed should be
fixed to the same number.}

\item{burn.in}{burn-in period of the MCMC sampling scheme; default is 10000.}

\item{num.it}{total number of MCMC iterations including burn-in; default is
40000.}

\item{summary}{logical; if \code{TRUE}, LBLcac will return a summary of the analysis. If \code{FALSE}, LBLcac will return the posterior samples for all parameters.}

\item{e}{a (small) number \eqn{\epsilon} in the null hypothesis of no association,
\eqn{H_0: |\beta| \le \epsilon}. The default is 0.1. Changing e from default of 0.1 may need choosing a
different threshold for Bayes Factor (one of the outputs) to infer
association. Only used if \code{summary = TRUE}.}

\item{ci.level}{Confidence level. The probability that the true value of \eqn{beta} will
be within the confidence interval. Default is 0.95 which corresponds to a 95\% posterior confidence interval. Only used if \code{summary = TRUE}.}
}
\value{
If \code{summary = FALSE}, return a list with the following components:
\describe{
\item{haplotypes}{The list of haplotypes used in the analysis. The last column is the reference haplotype.}

\item{beta}{Posterior samples of betas stored in a matrix.}

\item{lambda}{a vector of (num.it-burn.in) posterior samples of lambda.}

\item{freq}{Posterior samples of the frequency of haplotypes stored in a matrix format, in the same order as haplotypes.}

\item{init.freq}{The haplotype distribution used to initiate the MCMC.}

}

If \code{summary = TRUE}, return the result of LBL_summary. For details, see \code{?LBL_summary}.
}
\description{
\code{LBLcac} is a Bayesian LASSO method developed to detect association between
common/rare haplotypes and dichtomous disease phenotype, based on MCMC algorithm.
The "cac" in the title stands for "case and control", as this function will handle
independent cases and controls study design. For other types of study designs, see \code{\link{LBLfam}} and
\code{\link{LBLcombined}}. This function takes standard pedigree format as input with an individual's
genotypes, phenotype and familiar relationships, uses
an algorithm developed by Burkett \emph{et al}. to impute possible haplotype configurations
(for more detail, see \href{https://www.rdocumentation.org/packages/hapassoc/versions/0.5-1/topics/pre.hapassoc}{pre.hapassoc} from \pkg{hapassoc} package).
The input does not allow missing observations, and therefore subjects with
missing data are removed. This fucntion returns an object containing summary statistics
for all haplotypes present in the data.
}
\examples{
 data(cac)
 LBLcac(cac)


}
\references{
Biswas S, Lin S (2012). Logistic Bayesian LASSO for Identifying
  Association with Rare Haplotypes and Application to Age-related Macular
  Degeneration. Biometrics, 68(2): 587-97.

Burkett, K., Graham, k. and McNeney, B. (2006). "hapassoc: Software for likelihood
inference of trait associations with SNP haplotypes and other attributes."
Journal of Statistical Software 16(2): 1-19.
}
\seealso{
\code{\link{LBLfam}}, \code{\link{LBLcombined}}, \code{\link{LBL_summary}}.
}
\author{
{Swati Biswas, Meng Wang, Yuan Zhang, Swati Biswas, Shili Lin}
}
