% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_LBL_summary.R
\name{print_LBL_summary}
\alias{print_LBL_summary}
\title{Posterior Inference for LBL}
\usage{
print_LBL_summary(LBL_summary_object)
}
\description{
\code{print_LBL_summary} Print LBL result summary
}
\references{
Biswas S, Lin S (2012). Logistic Bayesian LASSO for Identifying
  Association with Rare Haplotypes and Application to Age-related Macular
  Degeneration. Biometrics, 68(2): 587-97.
}
