% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QBLbs.R
\name{QBLbs}
\alias{QBLbs}
\title{Quantitative Bayesian Lasso for Detecting Rare (or Common) Haplotype Association Correcting for Population Stratification}
\usage{
QBLbs(
  dat,
  numSNPs = 5,
  allelic = TRUE,
  baseline = "missing",
  interaction = F,
  cov = 1,
  pooling.tol = 0,
  zero.tol = "missing",
  a = 20,
  b = 20,
  start.beta = 0.01,
  lambda = 1,
  D = 0,
  seed = NULL,
  e = 0.1,
  burn.in = 20000,
  num.it = 50000,
  sigmas = 1,
  asig = 2,
  bsig = 1,
  CC = 1000,
  dumpConvergence = F,
  n.chains = 1,
  plot.interval = NULL
)
}
\arguments{
\item{dat}{The data set should consist of n rows and 1+n.cov+2p (allelic format) or 1+n.cov+p (genotypic format) columns, 
where n is the number of individuals, n.cov is the number of covariates and PCs, 
and p is the number of SNPs. The first column is the trait, followed by covariates and PCs.
If in allelic format, the other 2*p columns are allels with one column for each allele of the single-locus genotypes.
If in genotypic format, the other p columns are genotypes with one column for each SNP.
Missing data are not allowed and will be removed.}

\item{numSNPs}{The number of SNPs which form the haplotypes.}

\item{allelic}{TRUE if in allelic format and FALSE if in genotypic format; default is TRUE.}

\item{baseline}{The name of the baseline haplotype (e.g., "h00000"); default is the one with the maximum frequency.}

\item{interaction}{TRUE if consider the interaction between covariates and SNPs; only considers interaction with a maximum of the first 2 covariates; FALSE if else.}

\item{cov}{The number of covariates and PCs.}

\item{pooling.tol}{Haplotypes whose frequency is below pooling tolerance will be pooled together; default is 0}

\item{zero.tol}{Tolerance for haplotype frequencies below which haplotypes are assumed not to exist; default is 1/(20*n) where n is the number of individuals.}

\item{a}{First hyperparameter of the Gamma(\eqn{a},\eqn{b}) prior for regression coefficients,
\strong{\eqn{\beta}}. The prior variance of \eqn{\beta} is 2/\eqn{\lambda^2}. The Gamma prior parameters a and b 
are formulated such that the mean and variance of
the Gamma distribution are \eqn{a/b} and \eqn{a/b^2}. The default value of a is 20.}

\item{b}{Second hyperparameter of the Gamma(a,b) distribution described above; default
is 20.}

\item{start.beta}{Starting value of all regression coefficients, \strong{\eqn{\beta}};
default is 0.01.}

\item{lambda}{Starting value of the \eqn{\lambda} parameter described above;
default is 1.}

\item{D}{Starting value of the D parameter, which is the within-population
inbreeding coefficient; default is 0.}

\item{seed}{Seed to be used for the MCMC in Bayesian Lasso; default is a
random seed. If exact same results need to be reproduced, seed should be
fixed to the same number.}

\item{e}{A (small) number \eqn{\epsilon} in the null hypothesis of no association,
\eqn{H_0: |\beta| \le \epsilon}. The default is 0.1. Changing e from the default of 0.1 may necessitate choosing a
different threshold for Bayes Factor (one of the outputs) to infer
association.}

\item{burn.in}{Burn-in period of the MCMC sampling scheme; default is 20000.}

\item{num.it}{Total number of MCMC iterations including burn-in; default is
50000.}

\item{sigmas}{Starting value of the \eqn{\sigma^2} parameter, which is the random error; default is 1.}

\item{asig}{First hyperparameter of the Inverse-Gamma(\eqn{asig}, \eqn{bsig}) prior for the random error,
\eqn{\sigma^2}. The prior mean and variance of \eqn{\sigma^2} is \eqn{bsig/(asig-1)} and 
\eqn{bsig^2/(asig-1)^2(asig-2)}, respectively. The default value of \eqn{asig} is 2.}

\item{bsig}{Second hyperparameter of Inverse-Gamma(\eqn{asig}, \eqn{bsig}) described above; default
is 1.}

\item{CC}{The constant value set for the proposal distribution for updating haplotype
frequencies, \strong{\eqn{f}}. The proposal distribution of the MH algorithm is 
Dirichlet\eqn{(a_1,...,a_p)} with \eqn{a_1+...+a_p=CC}; default is 1000.}

\item{dumpConvergence}{TRUE if check convergence; FALSE if else; default is FALSE.}

\item{n.chains}{The number of chains for checking convergence; default if 1.}

\item{plot.interval}{The interval between two posterior samples that are extracted to draw trace plots.}
}
\value{

}
\description{
\code{QBLbs} is designed to identify rare (or common) haplotype effects associated with a quantitative trait.
The software is able to adjust for population stratification using principal components (PCs). 
The input does not allow missing observations and subjects with missing data are removed. 
The function returns an object containing posterior inferences after the burn-in period.
}
