% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cac}
\alias{cac}
\title{An example file consisting of independent case/controls}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{column 1}{family ID, integers}
  \item{column 2}{individual ID, integers}
  \item{column 3}{father's ID. 0 for unknown father.}
  \item{column 4}{mother's ID. 0 for unknown mother.}
  \item{column 5}{Individual's sex. 1 = male and 2 = female.}
  \item{column 6}{Affection Status. 0 = unknown, 1 = unaffected and 2 = affected.}
  \item{column 7 - 16}{marker genotypes. Each marker is represented by two columns: one for each allele.}
}}
\usage{
cac
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
