% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBL_summary.R
\name{LBL_summary}
\alias{LBL_summary}
\title{Posterior Inference for LBL}
\usage{
LBL_summary(output, a, b, e = 0.1, ci.level = 0.95)
}
\arguments{
\item{output}{An object returned by LBL functions (\code{\link{LBL}}, \code{\link{famLBL}}, and \code{\link{cLBL}}) with `summary=FALSE`.}

\item{a}{First hyperparameter of the prior for regression coefficients,
\eqn{\beta}. The prior variance of \eqn{\beta} is 2/\eqn{\lambda^2} and \eqn{\lambda} has Gamma(a,b)
prior. The Gamma prior parameters a and b are such that the mean and variance of
the Gamma distribution are \eqn{a/b} and \eqn{a/b^2}.
The value will be transferred from LBL functions.}

\item{b}{Second hyperparameter of the Gamma(a,b) distribution described above.
The value will be transferred from LBL functions.}

\item{e}{A (small) number \eqn{\epsilon} in the null hypothesis of no association,
\eqn{H_0: |\beta| \le \epsilon}. The default is 0.1. Changing e from the default of 0.1 may necessitate choosing a
different threshold for Bayes Factor (one of the outputs) to infer
association.}

\item{ci.level}{Credible probability.. The probability that the true value of \eqn{beta} will
be within the credible interval. Default is 0.95 which corresponds to a 95\% posterior credible interval. Only used if `summary = TRUE`.}
}
\value{
A list with the following components:
\describe{
\item{haplotypes}{The list of haplotypes used in the analysis.}

\item{OR}{Posterior mean of odds ratio.}

\item{OR.CI}{95\% posterior credible sets for the ORs.}

\item{BF}{Bayes Factor estimates based on posterior samples.
 if the posterior samples are all greater than e, then BF is set to be 999.}

}
}
\description{
\code{LBL_summary} provides inferences based on the posterior samples. Specifically, this function will return posterior means,
credible intervals, and Bayes Factors (BF) estimates for the haplotypic effect coefficients.
}
