% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_LBL_summary.R
\name{print_LBL_summary}
\alias{print_LBL_summary}
\title{Posterior Inference for LBL}
\usage{
print_LBL_summary(LBL_summary_object)
}
\description{
Prints list result from LBL_summary function in a more legible fashion.
}
